requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.2
defusedxml==0.6.0
Django==3.1.6
django-allauth==0.44.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
docopt==0.6.2
gunicorn
idna==2.10
mysqlclient==2.0.3
oauthlib==3.1.0
pipreqs==0.4.10
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
twilio==6.32.0
urllib3==1.26.3
yarg==0.1.9
argon2-cffi==20.1.0
